Add files via upload
Write a program that computes the area of a geometric shape and determines if the geometric shape is
small, medium, or large. Whether Prompt is used to determine which geometric shape they want to compute the area of. There are
only two shapes “Rectangle” Or “Circle”. Based on the type of geometric shape print a message to the user that indicates their selection Prompt user to enter appropriate parameters to compute the area for the selected.